[CAS-161] Move extensions to their own package.
