+++ livedata/src/main/java/io/getstream/chat/android/livedata/entity/MessageEntity.kt
@@ -7,6 +7,7 @@ import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.Reaction
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.utils.SyncStatus
+import io.getstream.chat.android.livedata.extensions.getCid
 import java.util.*
 
 /**
