+++ livedata/src/main/java/io/getstream/chat/android/livedata/repository/ReactionRepository.kt
@@ -6,7 +6,7 @@ import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.utils.SyncStatus
 import io.getstream.chat.android.livedata.dao.ReactionDao
 import io.getstream.chat.android.livedata.entity.ReactionEntity
-import io.getstream.chat.android.livedata.isPermanent
+import io.getstream.chat.android.livedata.extensions.isPermanent
 import java.security.InvalidParameterException
 
 /**
