+++ livedata/src/test/java/io/getstream/chat/android/livedata/ChatErrorTest.kt
@@ -4,6 +4,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth
 import io.getstream.chat.android.client.errors.ChatNetworkError
 import io.getstream.chat.android.client.models.Message
+import io.getstream.chat.android.livedata.extensions.isPermanent
 import org.junit.Test
 import org.junit.runner.RunWith
